+++ EssentialApp/EssentialAppTests/FeedLoaderWithFallbackCompositeTests.swift
@@ -38,8 +38,8 @@ final class FeedLoaderWithFallbackCompositeTests: XCTestCase {
 	// MARK: - Helpers
 	
 	private func makeSUT(primaryResult: FeedLoader.Result, fallbackResult: FeedLoader.Result, file: StaticString = #file, line: UInt = #line) -> FeedLoader {
-		let primaryLoader = LoaderStub(result: primaryResult)
-		let fallbackLoader = LoaderStub(result: fallbackResult)
+		let primaryLoader = FeedLoaderStub(result: primaryResult)
+		let fallbackLoader = FeedLoaderStub(result: fallbackResult)
 		let sut = FeedLoaderWithFallbackComposite(primary: primaryLoader, fallback: fallbackLoader)
 		trackForMemoryLeaks(primaryLoader, file: file, line: line)
 		trackForMemoryLeaks(fallbackLoader, file: file, line: line)
@@ -72,17 +72,4 @@ final class FeedLoaderWithFallbackCompositeTests: XCTestCase {
 	private func uniqueFeed() -> [FeedImage] {
         [FeedImage(id: UUID(), description: "any", location: "any", url: anyURL())]
 	}
-
-    private class LoaderStub: FeedLoader {
-        
-        private let result: FeedLoader.Result
-
-        init(result: FeedLoader.Result) {
-            self.result = result
-        }
-
-        func load(completion: @escaping (FeedLoader.Result) -> Void) {
-            completion(result)
-        }
-    }
 }
